requirements.txt
paho_mqtt>=2.0.0
pyserial>=3.5
PyYAML>=6.0.1